refactor(within-py): break import cycles, collapse to a 3-layer DAG
The PyO3 bridge had two module import cycles (legal in Rust, but a
coupling/altitude smell): convert <-> results and config <-> objects.
Each was thin (one symbol crossing per direction), so each broke with a
single relocation, after which the module graph is a clean acyclic DAG.

- convert <-> results: move run_solve/run_batch into results (they build
  the Py result types). convert is now a pure leaf; results -> convert.
- config <-> objects: move PyPreconditioner into config, beside the rest
  of the preconditioner surface (PrecondInput, resolve_precond_input).
  objects -> config only.
- objects then held only PySolver, splitting the public Python entry
  points by form (class vs free fn) rather than responsibility. Merge it
  into api and delete objects.rs.

No public/Python API change: same #[pyclass(name=...)] names and the
same _within registrations. 96/96 Python tests pass; fmt + clippy clean.

Resulting layers: convert (leaf) -> results / config -> api -> lib.
